+++ packages/core/src/Node.ts
@@ -549,7 +549,7 @@ export class Node<Options = any, Storage = any> {
 
     this.name = this.config.name
 
-    if (config.defaultOptions) {
+    if (config.defaultOptions && Object.keys(config.defaultOptions).length > 0) {
       console.warn(
         `[tiptap warn]: BREAKING CHANGE: "defaultOptions" is deprecated. Please use "addOptions" instead. Found in extension: "${this.name}".`,
       )
@@ -598,7 +598,7 @@ export class Node<Options = any, Storage = any> {
   extend<ExtendedOptions = Options, ExtendedStorage = Storage>(
     extendedConfig: Partial<NodeConfig<ExtendedOptions, ExtendedStorage>> = {},
   ) {
-    const extension = new Node<ExtendedOptions, ExtendedStorage>(extendedConfig)
+    const extension = new Node<ExtendedOptions, ExtendedStorage>({ ...this.config, ...extendedConfig })
 
     extension.parent = this