+++ packages/core/src/PasteRule.ts
@@ -21,7 +21,7 @@ export type PasteRuleMatch = {
   data?: Record<string, any>
 }
 
-export type PasteRuleFinder = RegExp | ((text: string) => PasteRuleMatch[] | null | undefined)
+export type PasteRuleFinder = RegExp | ((text: string, event?: ClipboardEvent) => PasteRuleMatch[] | null | undefined)
 
 export class PasteRule {
   find: PasteRuleFinder
@@ -33,17 +33,21 @@ export class PasteRule {
     commands: SingleCommands
     chain: () => ChainedCommands
     can: () => CanCommands
+    pasteEvent: ClipboardEvent
+    dropEvent: DragEvent
   }) => void | null
 
   constructor(config: {
     find: PasteRuleFinder
     handler: (props: {
-      state: EditorState
-      range: Range
-      match: ExtendedRegExpMatchArray
-      commands: SingleCommands
-      chain: () => ChainedCommands
       can: () => CanCommands
+      chain: () => ChainedCommands
+      commands: SingleCommands
+      dropEvent: DragEvent
+      match: ExtendedRegExpMatchArray
+      pasteEvent: ClipboardEvent
+      range: Range
+      state: EditorState
     }) => void | null
   }) {
     this.find = config.find
@@ -54,12 +58,13 @@ export class PasteRule {
 const pasteRuleMatcherHandler = (
   text: string,
   find: PasteRuleFinder,
+  event?: ClipboardEvent,
 ): ExtendedRegExpMatchArray[] => {
   if (isRegExp(find)) {
     return [...text.matchAll(find)]
   }
 
-  const matches = find(text)
+  const matches = find(text, event)
 
   if (!matches) {
     return []
@@ -92,9 +97,11 @@ function run(config: {
   from: number
   to: number
   rule: PasteRule
+  pasteEvent: ClipboardEvent
+  dropEvent: DragEvent
 }): boolean {
   const {
-    editor, state, from, to, rule,
+    editor, state, from, to, rule, pasteEvent, dropEvent,
   } = config
 
   const { commands, chain, can } = new CommandManager({
@@ -113,7 +120,7 @@ function run(config: {
     const resolvedTo = Math.min(to, pos + node.content.size)
     const textToMatch = node.textBetween(resolvedFrom - pos, resolvedTo - pos, undefined, '\ufffc')
 
-    const matches = pasteRuleMatcherHandler(textToMatch, rule.find)
+    const matches = pasteRuleMatcherHandler(textToMatch, rule.find, pasteEvent)
 
     matches.forEach(match => {
       if (match.index === undefined) {
@@ -134,6 +141,8 @@ function run(config: {
         commands,
         chain,
         can,
+        pasteEvent,
+        dropEvent,
       })
 
       handlers.push(handler)
@@ -155,6 +164,8 @@ export function pasteRulesPlugin(props: { editor: Editor; rules: PasteRule[] }):
   let dragSourceElement: Element | null = null
   let isPastedFromProseMirror = false
   let isDroppedFromProseMirror = false
+  let pasteEvent = new ClipboardEvent('paste')
+  let dropEvent = new DragEvent('drop')
 
   const plugins = rules.map(rule => {
     return new Plugin({
@@ -177,14 +188,17 @@ export function pasteRulesPlugin(props: { editor: Editor; rules: PasteRule[] }):
 
       props: {
         handleDOMEvents: {
-          drop: view => {
+          drop: (view, event: Event) => {
             isDroppedFromProseMirror = dragSourceElement === view.dom.parentElement
+            dropEvent = event as DragEvent
 
             return false
           },
 
-          paste: (view, event: Event) => {
+          paste: (_view, event: Event) => {
             const html = (event as ClipboardEvent).clipboardData?.getData('text/html')
+
+            pasteEvent = event as ClipboardEvent
 
             isPastedFromProseMirror = !!html?.includes('data-pm-slice')
 
@@ -224,12 +238,17 @@ export function pasteRulesPlugin(props: { editor: Editor; rules: PasteRule[] }):
           from: Math.max(from - 1, 0),
           to: to.b - 1,
           rule,
+          pasteEvent,
+          dropEvent,
         })
 
         // stop if there are no changes
         if (!handler || !tr.steps.length) {
           return
         }
+
+        dropEvent = new DragEvent('drop')
+        pasteEvent = new ClipboardEvent('paste')
 
         return tr
       },